add examples + update funcx-endpoint version

+++ examples/losses/mse_loss.py
@@ -1,4 +1,4 @@
 import torch
 
 def mse_loss(pred, y):
-    return torch.nn.MSELoss()(pred, y.unsqueeze(-1))
+    return torch.nn.MSELoss()(pred.float(), y.unsqueeze(-1).float())
